# server/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Food Delivery API",
    "description": "API for food delivery system including user, vendor, delivery agent, and analytics functionality",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5050/api/v1",
      "description": "Local development server"
    }
  ],
  "paths": {
...
   "/vendor/createMenuItem": {
  "post": {
    "tags": ["Vendor"],
    "summary": "Create menu item",
    "operationId": "createMenuItem",
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "requestBody": {
      "content": {
        "multipart/form-data": {
          "schema": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "Breakfast"
              },
              "name": {
                "type": "string",
                "example": "tea"
              },
              "price": {
                "type": "integer",
                "example": 120
              },
              "category": {
                "type": "string",
                "example": "Veg"
              },
              "cuisine": {
                "type": "string",
                "example": "North Indian"
              },
              "image": {
                "type": "string",
                "format": "binary"
              }
            },
            "required": ["type", "name", "price", "category", "cuisine", "image"]
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Menu item created successfully",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "Menu item added successfully!"
                },
                "menuItem": {
                  "type": "object",
                  "properties": {
                    "vendorId": {
                      "type": "string",
                      "example": "68132e745ba22849c6499ded"
                    },
                    "type": {
                      "type": "string",
                      "example": "Breakfast"
                    },
                    "name": {
                      "type": "string",
                      "example": "tea"
                    },
                    "price": {
                      "type": "integer",
                      "example": 120
                    },
                    "image": {
                      "type": "string",
                      "example": "/item/1746615955872.webp"
                    },
                    "category": {
                      "type": "string",
                      "example": "Veg"
                    },
                    "cuisine": {
                      "type": "string",
                      "example": "North Indian"
                    },
                    "_id": {
                      "type": "string",
                      "example": "681b3e8eaf0bef6e777cd46d"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-05-07T11:05:50.703Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-05-07T11:05:50.703Z"
                    },
                    "__v": {
                      "type": "integer",
                      "example": 0
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
},
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}